+++ sparv/core/registry.py
@@ -15,7 +15,7 @@ from sparv.core import paths
 from sparv.core.console import console
 from sparv.core.misc import SparvErrorMessage
 from sparv.api.classes import (BaseOutput, Config, Export, ExportAnnotations, ExportAnnotationsAllSourceFiles,
-                               SourceAnnotations, SourceStructureParser, ModelOutput, OutputCommonData, Wildcard)
+                               SourceAnnotations, SourceStructureParser, ModelOutput, OutputMarker, Wildcard)
 
 modules_path = ".".join(("sparv", paths.modules_dir))
 core_modules_path = ".".join(("sparv", paths.core_modules_dir))
@@ -325,7 +325,7 @@ def _add_to_registry(annotator):
 
     for _param, val in inspect.signature(annotator["function"]).parameters.items():
         if isinstance(val.default, BaseOutput):
-            if not has_marker and val.annotation == OutputCommonData:
+            if not has_marker and val.annotation == OutputMarker:
                 has_marker = True
             ann = val.default
             cls = val.default.cls
@@ -390,8 +390,8 @@ def _add_to_registry(annotator):
                                         "The export subdirectory must include the module name as prefix.")
 
     if annotator["type"] in (Annotator.installer, Annotator.uninstaller) and not has_marker:
-        raise SparvErrorMessage(f"'{rule_name}' creates no OutputCommonData marker, which is required by all "
-                                "installers and uninstallers.")
+        raise SparvErrorMessage(f"'{rule_name}' creates no OutputMarker, which is required by all installers and "
+                                "uninstallers.")
 
     if module_name not in modules:
         modules[module_name] = Module(module_name)