+++ sparv/modules/xml_export/xml_utils.py
@@ -182,4 +182,4 @@ def install_compressed_xml(corpus, bz2file, out, export_path, host):
     filename = corpus + ".xml.bz2"
     remote_file_path = os.path.join(export_path, filename)
     util.install.install_path(bz2file, host, remote_file_path)
-    out.write("")
+    out.write()